search bar in nav bar and styling
