+++ frontend/src/components/splash/splash_content_container.js
@@ -1,0 +1,18 @@
+import { connect } from "react-redux";
+import {
+    fetchAllBeaches
+} from "../../actions/beach_actions";
+import SplashContent from "./splash_content";
+
+const mapStateToProps = state => ({
+    beaches: state.entities.beaches
+});
+
+const mapDispatchToProps = dispatch => ({
+    fetchAllBeaches: () => dispatch(fetchAllBeaches())
+})
+
+export default connect(
+    mapStateToProps, 
+    mapDispatchToProps
+)(SplashContent);
